first steps to validation
