+++ src/hakunapi-jsonfg/pom.xml
@@ -1,33 +1,48 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<parent>
-		<groupId>fi.nls.hakunapi</groupId>
-		<artifactId>src</artifactId>
-		<version>1.3.0-SNAPSHOT</version>
-	</parent>
-	<artifactId>hakunapi-jsonfg</artifactId>
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>fi.nls.hakunapi</groupId>
+        <artifactId>src</artifactId>
+        <version>1.3.0-SNAPSHOT</version>
+    </parent>
+    <artifactId>hakunapi-jsonfg</artifactId>
 
-	<dependencies>
-		<dependency>
-			<groupId>fi.nls.hakunapi</groupId>
-			<artifactId>hakunapi-core</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>fi.nls.hakunapi</groupId>
-			<artifactId>hakunapi-geojson</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>fi.nls.hakunapi</groupId>
-			<artifactId>hakunapi-proj-gt</artifactId>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<scope>test</scope>
-		</dependency>
-	</dependencies>
+    <dependencies>
+        <dependency>
+            <groupId>fi.nls.hakunapi</groupId>
+            <artifactId>hakunapi-core</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>fi.nls.hakunapi</groupId>
+            <artifactId>hakunapi-geojson</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>fi.nls.hakunapi</groupId>
+            <artifactId>hakunapi-proj-gt</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.fasterxml.jackson.core</groupId>
+            <artifactId>jackson-databind</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.networknt</groupId>
+            <artifactId>json-schema-validator</artifactId>
+            <version>1.0.88</version>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
 
 </project>
